+++ plugins/main/public/components/common/data-source/pattern/alerts/index.ts
@@ -1,3 +1,5 @@
-export * from './alerts-vulnerabilities';
 export * from './alerts-data-source-repository';
 export * from './alerts-data-source';
+export * from './vulnerabilities';
+export * from './mitre-attack';
+export * from './virustotal';